requirements.txt
flask
spacy
scikit-learn
pdfplumber
python-docx
numpy
